models/version: Move license code to dedicated module
