+++ src/models/version.rs
@@ -192,21 +192,6 @@ impl NewVersion {
     }
 }
 
-pub fn validate_license_expr(s: &str) -> AppResult<()> {
-    pub const PARSE_MODE: spdx::ParseMode = spdx::ParseMode {
-        allow_lower_case_operators: false,
-        allow_slash_as_or_operator: true,
-        allow_imprecise_license_names: false,
-        allow_postfix_plus_on_gpl: true,
-    };
-
-    spdx::Expression::parse_mode(s, PARSE_MODE).map_err(|_| {
-        cargo_err("unknown or invalid license expression; see http://opensource.org/licenses for options, and http://spdx.org/licenses/ for their identifiers")
-    })?;
-
-    Ok(())
-}
-
 fn strip_build_metadata(version: &str) -> &str {
     version
         .split_once('+')
@@ -216,7 +201,7 @@ fn strip_build_metadata(version: &str) -> &str {
 
 #[cfg(test)]
 mod tests {
-    use super::{validate_license_expr, TopVersions};
+    use super::TopVersions;
     use chrono::NaiveDateTime;
 
     #[track_caller]
@@ -285,19 +270,5 @@ mod tests {
                 newest: Some(version("1.0.4")),
             }
         );
-    }
-
-    #[test]
-    fn licenses() {
-        assert_ok!(validate_license_expr("MIT"));
-        assert_ok!(validate_license_expr("MIT OR Apache-2.0"));
-        assert_ok!(validate_license_expr("MIT/Apache-2.0"));
-        assert_ok!(validate_license_expr("MIT AND Apache-2.0"));
-        assert_ok!(validate_license_expr("MIT OR (Apache-2.0 AND MIT)"));
-        assert_ok!(validate_license_expr("GPL-3.0+"));
-
-        let error = assert_err!(validate_license_expr("apache 2.0"));
-        let error = format!("{error}");
-        assert!(error.starts_with("unknown or invalid license expression; see http"));
     }
 }
